The rsm command was failing with an AttributeError.
The proximate cause is that if config is read from a file (and the confih parameter is None)
then the config reference never gets initialized sensibly.

To fix, we replace all references to config with config_instance.

(It might have been simpler to assign config_instance to config, but
this seems unclean).

Signed-off-by: Jon Seymour <jon.seymour@wildducktheories.com>